requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-WTF==1.1.1
openpyxl==3.1.2
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==10.1.0
APScheduler==3.10.4
qrcode[pil]==8.0
marshmallow==3.21.0
waitress==3.0.0
flask-limiter==4.1.1
flask-talisman==1.1.0
flask-compress==1.15
